Move response body viewer to its own component to share with timeline
