+++ playground/src/components/ResponseBody/ResponseBodyBinary.tsx
