+++ playground/src/components/ResponseBody/index.ts
@@ -1,3 +1,4 @@
 export * from "./ResponseBody";
 export * from "./ResponseBodyText";
 export * from "./FailedRequest";
+export { TextOrJsonViewer } from "./TextJsonViewer";
