+++ playground/src/routes/traces.$traceId.tsx
